backend/requirements.txt
# Backend dependencies (Python 3.13 compatible)
fastapi>=0.115.0
uvicorn>=0.32.0
mysql-connector-python>=9.1.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
python-jose>=3.3.0
pycryptodome>=3.21.0
passlib>=1.7.4
python-multipart>=0.0.12
python-dotenv>=1.0.1

